processor/requirements.txt
google-cloud-storage
google-cloud-pubsub
